SDC-6528. Oracle CDC - Add option to buffer records in SDC memory.
tl;dr: Add an option to read all changes from specified tables and buffer in SDC memory. On seeing
commit, push it into the pipeline.

Long Version:
When we start LogMiner, we use a start date/scn and fixed time windows for each session. This
did reduce the resource utilization on Oracle to a certain extent, but there was still quite
a bit of PGA memory usage.

On investigation, the reason seems to be that Oracle actually
materializes everything from the redo logs, even if the WHERE clause in the SELECT query
actually excludes some data. Since we were setting `COMMITTED_DATA_ONLY` flag, LogMiner was
buffering commits from all tables in PGA memory, even if those tables were not specified in
the SELECT query.

To avoid this, now we provide the option to buffer in SDC memory instead. By doing this, we
avoid using PGA memory but use SDC memory instead. The advantage is that we only ever buffer
the data that is to be sent down the pipeline (on commit, on rollback it has to be discarded).
The following changes are included:

* New offset format: v3::commit_timestamp::commit_scn::commit_seq. This allows us to restart
  from the earliest timestamp where there could be unread data (commit_timestamp minus txn_window).
  When a commit is seen, the offset is updated to show which commit was last seen and how many
  records from that transaction were already sent down the pipeline. This allows us to restart
  from where we left off.

* Buffer: We buffer all statements for a transaction involving the required tables. Once the commit
  is seen we send it down the pipeline. To avoid potential duplicates due to restarting from beginning
  of window we use RS_ID and SSN to check if a statement was seen before, which is checked using a new
  data structure: `HashQueue` which is a queue which we can check for the presence of the item.

* On rollback or if we see an earlier commit(like at time of restart), we remove all buffered statements

On testing, performance seems to be equal or better and no memory pressure on Oracle. In some cases,
transactions which were too huge for the old code can now be read pretty easily. We must note in the docs
the high memory usage though.

Currently, a pipeline previously started without in-memory buffering cannot be changed to buffer data
in-memory and vice versa.

Change-Id: Ibc50e1d40dc29f772fa46062d374bfe540c773c9